added a total count display and option for the experiment views
git-svn-id: svn://orca.physics.unc.edu/Orca@2904 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Experiments/Katrin/KatrinDetectorView.m
@@ -67,7 +67,7 @@
 	[super drawRect:rect];
 }
 
-- (NSColor*) getColorForSet:(int)setIndex value:(int)aValue
+- (NSColor*) getColorForSet:(int)setIndex value:(unsigned long)aValue
 {
 	if(setIndex==0)return [focalPlaneColorScale getColorForValue:aValue];
 	else return [vetoColorScale getColorForValue:aValue];
