+++ Source/Widgets/Bar Graph/ORColorScale.m
@@ -135,7 +135,7 @@
 	[self setNeedsDisplay:YES];
 }
 
-- (NSColor*) getColorForValue:(unsigned short)aValue
+- (NSColor*) getColorForValue:(unsigned long)aValue
 {	
 	float h=[self bounds].size.height;
 	float w=[self bounds].size.width;
@@ -154,7 +154,7 @@
     return [colors objectAtIndex:index];
 }
 
-- (unsigned short) getFastColorIndexForValue:(unsigned short)aValue log:(BOOL)aLog integer:(BOOL)aInt minPad:(double)aMinPad
+- (unsigned short) getFastColorIndexForValue:(unsigned long)aValue log:(BOOL)aLog integer:(BOOL)aInt minPad:(double)aMinPad
 {	 
 
 	aValue = [colorAxis getPixAbsFast:aValue log:aLog integer:aInt minPad:aMinPad];
@@ -170,7 +170,7 @@
 }
 
 
-- (unsigned short) getColorIndexForValue:(unsigned short)aValue 
+- (unsigned short) getColorIndexForValue:(unsigned long)aValue 
 {	
 	aValue = [colorAxis getPixAbs:aValue];
 	if(aValue <= 0)return 0;
